Reland: Change view_model.h to use size_t.
This relands change
https://chromium-review.googlesource.com/c/chromium/src/+/3710687 .

Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3723670
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1017750}

+++ ash/app_list/app_list_presenter_unittest.cc
@@ -889,7 +889,7 @@ TEST_P(ProductivityLauncherTest, FolderItemViewNotAnimatingAfterClosingFolder) {
   EXPECT_EQ(original_folder_item_bounds, folder_item_view->GetBoundsInScreen());
 
   // No item layers are expected to be created.
-  for (int i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
+  for (size_t i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
     views::View* item_view = apps_grid_view_->view_model()->view_at(i);
     EXPECT_FALSE(item_view->layer()) << "at " << i;
   }
@@ -1004,7 +1004,7 @@ TEST_P(ProductivityLauncherTest,
             apps_grid_view_->GetItemViewAt(2)->GetBoundsInScreen());
 
   // Verify that item view layers have been deleted.
-  for (int i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
+  for (size_t i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
     views::View* item_view = apps_grid_view_->view_model()->view_at(i);
     EXPECT_FALSE(item_view->layer()) << "at " << i;
   }
@@ -1123,7 +1123,7 @@ TEST_P(ProductivityLauncherTest,
             apps_grid_view_->GetItemViewAt(2)->GetBoundsInScreen());
 
   // Verify that item view layers have been deleted.
-  for (int i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
+  for (size_t i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
     views::View* item_view = apps_grid_view_->view_model()->view_at(i);
     EXPECT_FALSE(item_view->layer()) << "at " << i;
   }
@@ -1242,7 +1242,7 @@ TEST_P(ProductivityLauncherTest,
             apps_grid_view_->GetItemViewAt(3)->GetBoundsInScreen());
 
   // Verify that item view layers have been deleted.
-  for (int i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
+  for (size_t i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
     views::View* item_view = apps_grid_view_->view_model()->view_at(i);
     EXPECT_FALSE(item_view->layer()) << "at " << i;
   }
@@ -1325,7 +1325,7 @@ TEST_P(ProductivityLauncherTest, ReorderedFolderItemDeletionDuringFolderClose) {
             apps_grid_view_->GetItemViewAt(3)->GetBoundsInScreen());
 
   // Verify that item view layers have been deleted.
-  for (int i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
+  for (size_t i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
     views::View* item_view = apps_grid_view_->view_model()->view_at(i);
     EXPECT_FALSE(item_view->layer()) << "at " << i;
   }
@@ -1423,7 +1423,7 @@ TEST_P(ProductivityLauncherTest,
             apps_grid_view_->GetItemViewAt(3)->GetBoundsInScreen());
 
   // Verify that item view layers have been deleted.
-  for (int i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
+  for (size_t i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
     views::View* item_view = apps_grid_view_->view_model()->view_at(i);
     EXPECT_FALSE(item_view->layer()) << "at " << i;
   }
@@ -1520,7 +1520,7 @@ TEST_P(ProductivityLauncherTest,
             apps_grid_view_->GetItemViewAt(3)->GetBoundsInScreen());
 
   // Verify that item view layers have been deleted.
-  for (int i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
+  for (size_t i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
     views::View* item_view = apps_grid_view_->view_model()->view_at(i);
     EXPECT_FALSE(item_view->layer()) << "at " << i;
   }
@@ -2970,7 +2970,7 @@ TEST_P(PopulatedAppListTest, RemoveFolderItemAfterFolderCreation) {
   // Verify that item layers have been destroyed after the drag operation ended.
   apps_grid_test_api_->WaitForItemMoveAnimationDone();
 
-  for (int i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
+  for (size_t i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
     views::View* item_view = apps_grid_view_->view_model()->view_at(i);
     EXPECT_FALSE(item_view->layer()) << "at " << i;
   }
@@ -2985,7 +2985,7 @@ TEST_P(PopulatedAppListTest, RemoveFolderItemAfterFolderCreation) {
   EXPECT_EQ(expected_folder_item_view_bounds,
             folder_item_view->GetBoundsInScreen());
   EXPECT_TRUE(AppListIsInFolderView());
-  for (int i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
+  for (size_t i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
     views::View* item_view = apps_grid_view_->view_model()->view_at(i);
     EXPECT_FALSE(item_view->layer()) << "at " << i;
   }
@@ -3044,7 +3044,7 @@ TEST_P(PopulatedAppListTest, ReparentLastFolderItemAfterFolderCreation) {
   // Verify that item layers have been destroyed after the drag operation ended.
   apps_grid_test_api_->WaitForItemMoveAnimationDone();
 
-  for (int i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
+  for (size_t i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
     views::View* item_view = apps_grid_view_->view_model()->view_at(i);
     EXPECT_FALSE(item_view->layer()) << "at " << i;
   }
@@ -3066,7 +3066,7 @@ TEST_P(PopulatedAppListTest, ReparentLastFolderItemAfterFolderCreation) {
   // Verify that item views have no layers after the folder has been opened.
   apps_grid_test_api_->WaitForItemMoveAnimationDone();
   EXPECT_TRUE(AppListIsInFolderView());
-  for (int i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
+  for (size_t i = 0; i < apps_grid_view_->view_model()->view_size(); ++i) {
     views::View* item_view = apps_grid_view_->view_model()->view_at(i);
     EXPECT_FALSE(item_view->layer()) << "at " << i;
   }
@@ -5966,7 +5966,7 @@ TEST_P(AppListPresenterHomeLauncherTest, LayerOnSecondPage) {
   // of the layers, including items on the second page.
   generator->MoveMouseTo(start_point.x(), 0);
   generator->ReleaseLeftButton();
-  for (int i = 0; i < apps_grid_view->view_model()->view_size(); ++i) {
+  for (size_t i = 0; i < apps_grid_view->view_model()->view_size(); ++i) {
     views::View* item_view = apps_grid_view->view_model()->view_at(i);
     EXPECT_FALSE(item_view->layer()) << "at " << i;
   }
